backend/pyproject.toml
[project]
name = "smartlife-agent-backend"
version = "1.0.0"
description = "SmartLife Agent Backend with MCP Integration"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "pydantic[email]>=2.5.0",
    "google-generativeai>=0.3.2",
    "bcrypt>=4.1.2",
    "PyJWT>=2.8.0",
    "sqlalchemy>=2.0.23",
    "python-dotenv>=1.0.0",
    "google-genai>=1.45.0",
    "mcp>=1.18.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
    "google-api-python-client>=2.108.0",
    "google-auth>=2.41.1",
    "pytz>=2024.1",
]
